+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-20T05:21:44.691Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-20T09:56:39.645Z")
     r"""The date and time the resource was last updated"""