+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K80129A933MRBY6BBSH3QWKY"
+        "01K80GSNR928CQJE76NYANQ57D"
     )
     r"""The id of the resource"""