+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K80129ABK01GAXB3E817A34M"
+        "01K80GSNRCJWBJAKJ6BQW8MXYN"
     )
     r"""The id of the resource"""